requirements.txt
qdrant-client[fastembed]>=1.17.1
python-dotenv>=1.0.1
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
